Floating Action Button: shrink button when pushed.
You're applying a force into the screen when you press the button, so it should get smaller. Also all of the faces fly out ot it, so it should get smaller when they're outside, too. I think some of Material Design has "push down upon" mapping to "pick up" which is also useful in some situations. For a button having it shrink when pushed feels much more tactile.