constraints: pretend constraint system.
This is my first attempt at a constraint system (not even a solver) and an example to implement crude touch scrolling. There are some tricky bits even in a simple constraint system like scrolling (the constraints are don't scroll off the top and don't scroll off the bottom). I want the constraint to apply deltas rather than the manipulated value doing it.