constraints: twitter: start on expansion mechanic.
Super buggy. I need to wrap the solver so that I can manage edit sessions when there are multiple edits going on at once. If you drag something while a different animation is going on then it gets messed up. I don't think I can keep the bottom sticky constraint, either. I'm not sure what to do about that. Maybe instead of an opened tweet staying inline it should immediately animate to the top? This is a fun exploration!