constraints: start of reworked manipulator.
The manipulator is now significantly simplified. I recognized that there are (currently) three sources of motion for the manipulator: drag, velocity animation and constraint animation. We need to do something different for each when we hit a constraint, and the new code separates them so that we can cleanly do that. It needs debugging though, there are some discontinuities and bugs.