constraints: don't cancel a constraint animation when it reviolates the constraint it's attempting to satisfy.
A constraint may specify an underdamped spring or something like that, where it takes the variable in and out of violation but doesn't want to be interrupted or recreated while doing so. Two more obvious bugs remain:
 - if the velocity imparted by a drag would take a variable out of violation (i.e.: you're flicking out of overscroll) then we shouldn't constrain; we should just let it go.
 - when dragging into violation the overdrag coefficient (currently a hardcoded 0.5, but eventually coming from the motion constraint so that we can do hard limits, etc) we apply the coefficient to the whole delta, not just the violating part of the delta. This is easy to fix if the overdrag coefficient is always just a number and not some non-linear function.

Eventually I want to support non-linear functions for drag coefficients and then use that mechanism for doing something like the iMessage relative springy scrolling.