constraints: don't do constraint animation when the velocity from a drag would take the variable out of violation anyway.
This is the last obvious bug solved. All of the scrolling/frictiony things work pretty well now. I'm not sure if all of these rules will hold for a different simulation, like gravity.

Next I'm going to expand the definition of a motion constraint a bit so that they can be enforced hard (with optional rebound) or with springs. I should do some other examples, too, like the new mobile Google Maps pull up UI, which could be nicely expressed with constraints. With some more invention I hope to be able to express motion constraints that instead of preventing the motion of the variable that's being dragged instead respond to it. That will enable the iMessage style scrolling.