constraints: start of google maps result detail interaction.
Mobile Google Maps has a nice progressive exposure UI when dragging upwards on the result info at the bottom of the screen. This is a start at that. The basics are there, but it needs snapping on the manipulator so that it can't be left in intermediate positions and has some gravity. Interesting work to be done there!

The layout of the text content is also not complete, but that's not really the focus here. Some interesting details in here are the dimming shadow that appears at the top of the image and the small navbar when you scroll past the image. Those are hard to do in traditional layout systems (and hard in HTML/CSS) but quite straightforward with constraints.