constraints: start experiementing with pager/snapping behavior via animation constraints.
I haven't come up with a nice way of describing pager behavior yet, so the test case for this is a bit longer that I'd like.

I accomplished this by allowing constraints to only enforce on animations or when the finger was released. I think that the logic in manipulator.js for when to apply a motion constraint needs to move out to the actual constraint itself since there are some behavior variations. For example, in scrolling if the velocity will take you out of violation then you don't want to enforce the constraint (that is, if you're off the top of the list and flick downwards the view should scroll downwards, it shouldn't bounce around at the top; maybe actually the spring should be used and its velocity transferred to a friction animation once the scroll has returned to the valid range; I haven't tried that yet).

I'll update the gmaps example to use this kind of snapping next, then I'll break some of the constraint enforcement logic out of the manipulator (although I want to keep it so that the manipulator owns the constraint enforcement animation, etc).