constraints: properly implement captive vs non-captive motion constraints.
A captive constraint keeps the value once it's violated, where as a non-captive constraint releases the value as soon as it is out of violation. So you need to use captive constraints when you have an animation which bounces out of violation and back in again. Most of the time you don't want to use captive motion constraints, but they're really useful for pagers and other things where you're animating to a particular value.

Interestingly with this change I can no longer reproduce the old scroll behavior I had, because it would *not* capture the variable if the velocity would take it out of violation with friction, but it would capture it once it violated so I could have an underdamped spring. Now if the spring is underdamped then the value "bounces" off of the end of a scroll and starts scrolling back the other way if it's non-captive. If it is captive then flicking while in overdrag causes the value to just bounce around around that overdrag extent. For now I think this is OK, but perhaps I'll have to add a new option to bring back the possibility of the old behavior if it shows up again... It isn't an issue for any of the other examples.