constraints: checkpoint work to move various flags from ops to the constraint.
In particular, animationOnly has now been moved out of the "op" and is now a property of the motion constraint itself. I added a few others to try to maintain compatibility with the old demos, but it didn't really pan out. So I'm going to remove the other new flags and rewrite the GMaps example at least next.