constraints: eliminate the concept of an animationOnly constraint.
We can use the existing overdragCoefficient for this -- just set it to zero and then it doesn't change things while dragging. We now sort multiple violations on the overdragCoefficient in order to find the most constraining motion constraint to enforce.

This means that the only special flag we have is "captive" which I think is a good one. I think there might be two variants of "captive":
 - fully captive (currently implemented): the constraint enforces regardless of what's going on and doesn't relinquish control of the value regardless of the velocity.
 - one direction captive: only captures the value if there's no velocity or the velocity is going against the constraint; once captured it's the same as captive. This is like what we had before with underdamped end springs.