constraints: ios: add pull down menu as well as control center.
This was actually super interesting. The pull down menu needs a different physics model depending on which direction you pull it in. If you're sending it up then it has to accelerate upwards; if you're pulling down it has to accelerate downwards (and bounce a bit). I hadn't modelled a system like that with MotionConstraints before, so it's nice that it was fairly straightforward (although it could be improved).